+++ src/main/java/nameco/stikku/user/UserController.java
@@ -1,7 +1,10 @@
 package nameco.stikku.user;
 
+import jakarta.servlet.http.HttpServletRequest;
+import nameco.stikku.advice.exception.AccesDeniedException;
 import nameco.stikku.game.GameService;
 import nameco.stikku.game.dto.GameResponseDto;
+import nameco.stikku.resolver.Auth;
 import nameco.stikku.responseDto.MessageResponse;
 import nameco.stikku.user.dto.UserDTO;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -9,6 +12,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.nio.file.AccessDeniedException;
 import java.util.List;
 
 @RestController
@@ -36,12 +40,18 @@ public class UserController {
     }
 
     @PutMapping("/{user_id}")
-    public ResponseEntity<User> updateUser(@PathVariable("user_id") Long userId, @RequestBody UserDTO userDTO) {
+    public ResponseEntity<User> updateUser(@Auth String tokenUserId, @PathVariable("user_id") Long userId, @RequestBody UserDTO userDTO) {
+        if(!tokenUserId.equals(userId.toString())){
+            throw new AccesDeniedException("권한이 없습니다.");
+        }
         return new ResponseEntity<>(userService.updateUser(userId, userDTO), HttpStatus.OK);
     }
 
     @DeleteMapping("/{user_id}")
-    public ResponseEntity<MessageResponse> deleteUser(@PathVariable("user_id") Long userId) {
+    public ResponseEntity<MessageResponse> deleteUser(@Auth String tokenUserId, @PathVariable("user_id") Long userId) {
+        if(!tokenUserId.equals(userId.toString())){
+            throw new AccesDeniedException("권한이 없습니다.");
+        }
         String deletedUserId = userService.deleteUser(userId);
         return new ResponseEntity<>(new MessageResponse("User " + deletedUserId + " deleted successfully"), HttpStatus.OK);
     }
